устранение дефектов

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -7,7 +7,6 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.TreeSet;
 
 @Data
 public class User {
@@ -15,20 +14,20 @@ public class User {
     @Pattern(regexp = "\\S*$")
     private String login;
     private String name;
-    private int id;
+    private long id;
     @NotBlank(message = "email blank")
     @Email(message = "invalid email")
     private String email;
     @NotBlank
     @Pattern(regexp = "^\\d{4}-\\d{2}-\\d{2}$", message = "invalid format birthday")
     private String birthday;
-    private Set<Integer> friends = new HashSet<>();
+    private Set<Long> friends = new HashSet<>();
 
-    public void addFriend(int friendId) {
+    public void addFriend(long friendId) {
         friends.add(friendId);
     }
 
-    public void removeFriend(int friendId) {
+    public void removeFriend(long friendId) {
         friends.remove(friendId);
     }
 }
